🔧 fix: update API calls to use NextAuth session tokens - Add authActions for centralized authentication handling - Update all server actions to use NextAuth session instead of custom auth - Modify client-side components to work with new authentication flow - Update functions to handle NextAuth session tokens properly - Ensure consistent authentication across cart, orders, bookmarks, and reviews Ref #48
