+++ src/lib/functions/authFunctions.ts
@@ -1,6 +1,5 @@
 'use server';
 
-import { EmailCheckResult, NicknameCheckResult, SignUpRequestData, SignUpResult } from '@/app/(auth)/sign-up/_types';
 import { ApiRes } from '@/types/api.types';
 import { LoginCredentials, LoginResult } from '@/types/auth.types';
 import { User } from '@/types/user.types';
@@ -176,65 +175,8 @@ export async function logoutAction(): Promise<ApiRes<{ message: string }>> {
   }
 }
 
-/**
- * 회원가입 서버 액션
- * @param signUpData - 회원가입 데이터
- * @returns 표준화된 API 응답 형태
- */
-export async function signUpAction(signUpData: SignUpRequestData): Promise<SignUpResult> {
-  try {
-    const result = await signUpUser(signUpData);
 
-    // 회원가입 성공 시 결과만 반환
-    if (result.ok === 1) {
-      console.log('[회원가입 액션] 회원가입 성공');
-    }
 
-    return result;
-  } catch (error) {
-    console.error('[회원가입 액션] 오류:', error);
-    return {
-      ok: 0,
-      message: '회원가입 처리 중 오류가 발생했습니다.',
-    };
-  }
-}
-
-/**
- * 이메일 중복 확인 서버 액션
- * @param email - 확인할 이메일
- * @returns 이메일 사용 가능 여부
- */
-export async function checkEmailAction(email: string): Promise<EmailCheckResult> {
-  try {
-    const result = await checkEmailAvailability(email);
-    return result;
-  } catch (error) {
-    console.error('[이메일 확인 액션] 오류:', error);
-    return {
-      ok: 0,
-      message: '이메일 확인 중 오류가 발생했습니다.',
-    };
-  }
-}
-
-/**
- * 닉네임 중복 확인 서버 액션
- * @param nickname - 확인할 닉네임
- * @returns 닉네임 사용 가능 여부
- */
-export async function checkNicknameAction(nickname: string): Promise<NicknameCheckResult> {
-  try {
-    const result = await checkNicknameAvailability(nickname);
-    return result;
-  } catch (error) {
-    console.error('[닉네임 확인 액션] 오류:', error);
-    return {
-      ok: 0,
-      message: '닉네임 확인 중 오류가 발생했습니다.',
-    };
-  }
-}
 
 // ============================================================================
 // 내부 함수들 (서버 액션에서만 사용)
@@ -365,145 +307,5 @@ async function refreshAccessToken(refreshToken: string): Promise<ApiRes<{ access
   }
 }
 
-/**
- * 회원가입 API 호출 함수
- * @param signUpData - 회원가입 데이터
- * @returns 표준화된 API 응답 형태
- */
-async function signUpUser(signUpData: SignUpRequestData): Promise<SignUpResult> {
-  try {
-    const res = await fetch(`${API_URL}/users`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Client-Id': CLIENT_ID,
-      },
-      body: JSON.stringify(signUpData),
-    });
 
-    const data = await res.json();
 
-    if (!res.ok) {
-      console.error('[회원가입 함수] API 오류:', data);
-      return {
-        ok: 0,
-        message: data.message || '회원가입에 실패했습니다.',
-        errors: data.errors,
-      };
-    }
-
-    const user: User = data.item;
-    if (!user) {
-      return {
-        ok: 0,
-        message: '회원가입 정보가 올바르지 않습니다.',
-      };
-    }
-
-    console.log('[회원가입 함수] 성공:', { userId: user._id, email: user.email });
-    return {
-      ok: 1,
-      item: user,
-    };
-  } catch (error) {
-    console.error('[회원가입 함수] 네트워크 오류:', error);
-    return {
-      ok: 0,
-      message: '네트워크 오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
-    };
-  }
-}
-
-/**
- * 이메일 중복 확인 API 호출 함수
- * @param email - 확인할 이메일
- * @returns 이메일 사용 가능 여부
- */
-async function checkEmailAvailability(email: string): Promise<EmailCheckResult> {
-  console.log(`[checkEmailAvailability] API 요청: ${API_URL}/users/email?email=${email}`);
-  try {
-    const res = await fetch(`${API_URL}/users/email?email=${encodeURIComponent(email)}`, {
-      method: 'GET',
-      headers: {
-        'Client-Id': CLIENT_ID,
-      },
-    });
-
-    const data = await res.json();
-    console.log('[checkEmailAvailability] API 응답 상태:', res.status);
-    console.log('[checkEmailAvailability] API 응답 데이터:', data);
-
-    if (data.ok === 1) {
-      console.log('[checkEmailAvailability] 이메일 사용 가능');
-      return {
-        ok: 1,
-        item: { available: true },
-      };
-    } else if (data.ok === 0) {
-      console.log('[checkEmailAvailability] 이메일 이미 사용 중');
-      return {
-        ok: 1, // API 통신은 성공했으므로 ok: 1, 비즈니스 로직 결과는 available: false
-        item: { available: false },
-      };
-    } else {
-      console.error('[이메일 확인 함수] API 응답 형식 오류:', data);
-      return {
-        ok: 0,
-        message: data.message || '서버 응답이 올바르지 않습니다.',
-      };
-    }
-  } catch (error) {
-    console.error('[이메일 확인 함수] 네트워크 오류:', error);
-    return {
-      ok: 0,
-      message: '네트워크 오류가 발생했습니다.',
-    };
-  }
-}
-
-/**
- * 닉네임 중복 확인 API 호출 함수
- * @param nickname - 확인할 닉네임
- * @returns 닉네임 사용 가능 여부
- */
-async function checkNicknameAvailability(nickname: string): Promise<NicknameCheckResult> {
-  console.log(`[checkNicknameAvailability] API 요청: ${API_URL}/users/name?name=${nickname}`);
-  try {
-    const res = await fetch(`${API_URL}/users/name?name=${encodeURIComponent(nickname)}`, {
-      method: 'GET',
-      headers: {
-        'Client-Id': CLIENT_ID,
-      },
-    });
-
-    const data = await res.json();
-    console.log('[checkNicknameAvailability] API 응답 상태:', res.status);
-    console.log('[checkNicknameAvailability] API 응답 데이터:', data);
-
-    if (data.ok === 1) {
-      console.log('[checkNicknameAvailability] 닉네임 사용 가능');
-      return {
-        ok: 1,
-        item: { available: true },
-      };
-    } else if (data.ok === 0) {
-      console.log('[checkNicknameAvailability] 닉네임 이미 사용 중');
-      return {
-        ok: 1, // API 통신은 성공했으므로 ok: 1, 비즈니스 로직 결과는 available: false
-        item: { available: false },
-      };
-    } else {
-      console.error('[닉네임 확인 함수] API 응답 형식 오류:', data);
-      return {
-        ok: 0,
-        message: data.message || '서버 응답이 올바르지 않습니다.',
-      };
-    }
-  } catch (error) {
-    console.error('[닉네임 확인 함수] 네트워크 오류:', error);
-    return {
-      ok: 0,
-      message: '네트워크 오류가 발생했습니다.',
-    };
-  }
-}
